Decided since im making a time keeping app. the ability to test and manipulate the internal test time was important.
Future plans for this is to have an offset stored in the database so the running instance can have the ability to call the TestTimeService.

Added test tiem service to allow unit tests more freedom with the system time.

added many tests to it. Maybe move this into utility services of some kind.